Add Salesman Crud Implementation

+++ src/main/java/ccf/project/repository/ProductTypeRepository.java
@@ -4,10 +4,11 @@ import ccf.project.domain.ProductTypeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import java.util.List;
 
 @Repository
-public interface ProductTypeRepository extends JpaRepository<ProductTypeModel, Long> {
+public interface ProductTypeRepository extends JpaRepository<ProductTypeModel, Integer> {
 
     ProductTypeModel findByType(String type);
