+++ src/main/java/ccf/project/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserModel, Long> {
+public interface UserRepository extends JpaRepository<UserModel, Integer> {
     Optional<UserModel> findByUsername(String name);
 
     Integer deleteByUsername(String username);
